Move escapeData to lib
