+++ packages/frontend/lib/escapeData.tsx
