+++ packages/frontend/web/components/elements/TweetBlockComponent.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { css } from 'emotion';
 import { palette } from '@guardian/pasteup/palette';
 import { body } from '@guardian/pasteup/typography';
-import { unescapeData } from '@frontend/web/server/escapeData';
+import { unescapeData } from '@frontend/lib/escapeData';
 
 // fallback styling for when JS is disabled
 const noJSStyling = css`
